app/task/requirements.txt
django==5.0.1
djangorestframework==3.14.0
psycopg2>=2.8
